Change id_animal type from int to Long
Cambio de tipo de dato a entidad Animales

+++ src/main/java/Semovientes/app/model/Animales.java
@@ -6,7 +6,7 @@ import jakarta.persistence.*;
 public class Animales {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_animal;
+    private Long id_animal;
     
     @ManyToOne(fetch=FetchType.LAZY)
     @JoinColumn(name="finca_id")
